Former-commit-id: 246baac272eaad3557c10dd6337cfcff1df1882a [formerly a8bcabc2671f5656649fea0754d68bd34fadd061]
Former-commit-id: 02a8ffabeb66fd4df9754485965d0472db7caeed

+++ librapid/src/multiprecToString.cpp
@@ -10,7 +10,8 @@ namespace librapid {
 		auto res = val.get_str(exp, base);
 
 		if (exp > 0) {
-			if (exp >= res.length()) res += std::string(exp - res.length() + 1, '0');
+			if (static_cast<size_t>(exp) >= res.length())
+				res += std::string(static_cast<size_t>(exp) - res.length() + 1, '0');
 			res.insert(exp, ".");
 			return res;
 		} else {
